requirements.txt

numpy
opencv-python
easydict

torch>=1.7.1
pillow
tqdm

damei

wandb  # for UNet
